Backend/requirements.txt
asgiref
async-timeout
boto3
botocore
cachetools
certifi
charset-normalizer
Django
django-cors-headers
django-redis
django-storages
djangorestframework
djangorestframework-simplejwt
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
gunicorn
httplib2
idna
jmespath
Pillow
protobuf
psycopg2-binary
pyasn1
pyasn1-modules
PyJWT
pyparsing
python-dateutil
python-decouple
pytz
razorpay
redis
requests
rsa
s3transfer
six
sqlparse
uritemplate
urllib3
